LayoutTests/imported/w3c/web-platform-tests/webstorage/event_initstorageevent.window-expected.txt


PASS initStorageEvent with 0 arguments
FAIL initStorageEvent with 1 argument assert_equals: event.key expected (object) null but got (string) "undefined"
PASS initStorageEvent with 8 sensible arguments
FAIL initStorageEvent with 8 null arguments assert_equals: event.key expected (object) null but got (string) "null"
FAIL initStorageEvent with 8 undefined arguments assert_equals: event.key expected (object) null but got (string) "undefined"
